+++ misc/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## unreleased
 
+-   tflite: improve crop size error message #24
+
 ## `v0.2.10`
 
 -   tflite: fix sub-stream toggle